Refactor shared components location
