+++ src/components/layout/Footer.tsx
@@ -1,4 +1,4 @@
-import Title from "@/app/Title";
+import Title from "@/components/common/Title";
 
 export default function Footer() {
   return (
